Add files via upload
Approach
The approach of the code is as follows:

First, the code checks if the lengths of the two input strings, s and goal, are not equal or if the length of s is less than or equal to 1. If either of these conditions is true, the function returns false. This is because for the two strings to be buddy strings, they must have the same length and contain at least two differing characters.

Next, the code checks if s and goal are identical. If they are, it means that all characters are the same. In this case, the code creates an set, distinctChars, to store the distinct characters from s. If the size of distinctChars is less than the length of s, it means there are duplicate characters, and we can swap two of them to make s and goal equal. The function returns true in this case.

If the previous conditions are not met, it means that s and goal are different strings, and we need to check if they can be made equal by swapping exactly two characters.

The code iterates through each character of s and goal simultaneously, comparing the characters at each index. Whenever a pair of characters at the same index is different, the index is added to the index vector.

After iterating through both strings, the code checks if the size of index is exactly 2. If it's not, it means there are either more or fewer than two differing characters, and the strings cannot be made equal by swapping two characters. In this case, the function returns false.

If there are exactly two differing characters, the code proceeds to swap these characters in s by using the indices stored in the index vector.

Finally, the code checks if s is now equal to goal after the swap. If they are equal, it means that s and goal can be made equal by swapping two characters, and the function returns true. Otherwise, it means the swap did not result in s becoming equal to goal, and the function returns false.

Complexity
Time complexity: O(n)O(n)O(n)

Space complexity: O(n)O(n)O(n)