Update 712. Minimum ASCII Delete Sum for Two Strings.cpp
##Approach

The solution utilizes dynamic programming to find the minimum ASCII delete sum of s1 and s2. A 2D DP array dp is used to store the minimum delete sum for subproblems.

The recursive function solve is defined, which takes four parameters: i and j are the current indices for s1 and s2, respectively, and s1 and s2 are the input strings.

In the solve function, the following steps are followed:

a. Check if the value for the current state (i, j) is already computed and stored in the dp array. If yes, return the value to avoid recomputation.

b. If i reaches the end of s1, it means all characters of s1 have been processed. In this case, sum the ASCII values of the remaining characters in s2 (from index j to the end) and return the sum. These characters need to be deleted to make s1 and s2 equal.

c. If j reaches the end of s2, it means all characters of s2 have been processed. Similar to the previous case, sum the ASCII values of the remaining characters in s1 (from index i to the end) and return the sum.

d. If neither i nor j has reached the end of their respective strings, then two options are possible:

Calculate the sum of the ASCII value of the current character of s1 with the minimum delete sum of the remaining characters of s1 and s2 (by calling the solve function recursively with i+1 and j).

Calculate the sum of the ASCII value of the current character of s2 with the minimum delete sum of the remaining characters of s1 and s2 (by calling the solve function recursively with i and j+1).

If the current characters of s1 and s2 are equal, an additional option is available:

Calculate the minimum delete sum of the remaining characters of s1 and s2 (by calling the solve function recursively with i+1 and j+1) without including the current characters in the sum since they are already equal.
e. Return the minimum of the above three options as the minimum delete sum for the current state (i, j) and store it in the dp array for future reference.

Complexity
Time complexity:O(n1∗n2)O(n1*n2)O(n1∗n2)
Space complexity:O(n1∗n2)O(n1*n2)O(n1∗n2)
Code